Restructured the repository: Created a project for the ready wrappers. Made OuputDir independent of the directory the assembly is in.
