+++ QtSharp.sln
@@ -1,13 +1,15 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2013
-VisualStudioVersion = 12.0.30501.0
+VisualStudioVersion = 12.0.21005.1
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QtSharp", "QtSharp\QtSharp.csproj", "{861046E1-DA11-4229-B9B6-45426E7A723E}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QtSharp.CLI", "QtSharp.CLI\QtSharp.CLI.csproj", "{DBDEB3EE-2EB3-4F03-8276-A7D5A350EB95}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QtSharp.Tests", "QtSharp.Tests\QtSharp.Tests.csproj", "{5E8F503D-1B37-4298-855D-06F81B5B4630}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Qt5Core.Gen", "Qt5Core.Gen\Qt5Core.Gen.csproj", "{BACBCD61-F377-49CB-BD93-4FB539146B63}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -38,6 +40,12 @@ Global
 		{5E8F503D-1B37-4298-855D-06F81B5B4630}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5E8F503D-1B37-4298-855D-06F81B5B4630}.Release|Any CPU.Build.0 = Release|Any CPU
 		{5E8F503D-1B37-4298-855D-06F81B5B4630}.Release|x86.ActiveCfg = Release|Any CPU
+		{BACBCD61-F377-49CB-BD93-4FB539146B63}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BACBCD61-F377-49CB-BD93-4FB539146B63}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BACBCD61-F377-49CB-BD93-4FB539146B63}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{BACBCD61-F377-49CB-BD93-4FB539146B63}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BACBCD61-F377-49CB-BD93-4FB539146B63}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BACBCD61-F377-49CB-BD93-4FB539146B63}.Release|x86.ActiveCfg = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
