+++ QtSharp/ProcessHelper.cs
@@ -1,16 +1,18 @@
 ﻿using System;
+using System.IO;
 using System.Diagnostics;
 
 namespace QtSharp
 {
     public class ProcessHelper
     {
-        public static string Run(string path, string args, out string error)
+        public static string Run(string path, string args, out string error, string wdir="")
         {
             try
             {
                 using (Process process = new Process())
                 {
+                    process.StartInfo.WorkingDirectory = wdir;
                     process.StartInfo.FileName = path;
                     process.StartInfo.Arguments = args;
                     process.StartInfo.UseShellExecute = false;
