+++ QtSharp/QtSharp.cs
@@ -138,10 +138,16 @@ namespace QtSharp
                 RenameCasePattern.UpperCamelCase).VisitLibrary(driver.ASTContext);
         }
 
+        static public string GetOutputDir()
+        {
+            return Path.GetFullPath(Path.Combine(Directory.GetCurrentDirectory(), "../../..", "Qt5Core.Gen")); // The base directory of QtSharp project + Qt5Core.Gen
+        }
+
 		public void Setup(Driver driver)
 		{
 			driver.Options.GeneratorKind = GeneratorKind.CSharp;
 		    string qtModule = "Qt" + this.module;
+            driver.Options.OutputDir = GetOutputDir();
             driver.Options.addDefines("_CRTIMP=");
 		    driver.Options.MicrosoftMode = false;
             driver.Options.TargetTriple = this.target;
@@ -172,7 +178,7 @@ namespace QtSharp
             driver.Options.Libraries.Add(this.library);
 		    if (this.module == "Core")
 		    {
-                string dir = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
+                string dir = GetOutputDir();
                 driver.Options.CodeFiles.Add(Path.Combine(dir, "QEventArgs.cs"));
                 driver.Options.CodeFiles.Add(Path.Combine(dir, "QEventHandler.cs"));
                 driver.Options.CodeFiles.Add(Path.Combine(dir, "QObject.cs"));
